move functionality to separate file, add docs
